Provides a unique ID for entities already present in the pmird database.

pm_get_present is a helper function to define a unique ID for entities already present in a table in the pmird database. This function is used by pm_update_import.

pm_get_present(con, table_name)

Arguments

con
Either a database connection (see pm_update_import) or a data frame wit the same format as the table indicated by table_name in the pmird database (same column names, no additional columns).

table_name
See pm_update_import.

Value

A data frame with the same rows as the table table_name in the pmird database (if con is a database connection) or in the provided data frame (id x is a data frame) and with the folowing two columns:

The ID (primary key).

A column x_present with the unique ID which can be used to match entities across data frames.